-Changed table to 2x2 sprite -Crafting table -Pickups go through walls as long as the player's magnet is pulling them -Much better day night cycle -Added clock -You can now save day night cycle -Limited speed of camera movement -Functional returning to spawn -Added functional crafting -Separated itemData properties so you can open and close them in inspector -Added collider generation for items -Basic combat -Easy to implement sounds -Tree sounds -Damage Value property for player -Added visual for breaking a block -Cleaner code in general -Inheritance tree for entitiies and enemies -Inheritance tree for inventory slots and crafting slots -No longer storing trees -Much better way for chopping trees -Added 18x18 dirt wall -Enemy spawner -Enemies spawn on top of groun instead of inside ground -Enemies take knockback -Crafting Slot -Selectable Slot -Slot UI -Bug fixes

+++ Assets/Scripts/Data/ToolData.cs
@@ -14,5 +14,5 @@ public class ToolData
 {
     public byte worldLayer;
     public Tool toolType;
-    public float strength;
+    public int strength;
 }
